Enable multi-channel/multi-device subdevice mappings, and improve device-related XML error checking

Closes #1333 and #894

See merge request opencpi/opencpi!521
(cherry picked from commit 543025c5cd0b1b2bc5e180ea7ea4ecc7646584dd)

Conflicts:
	doc/odt/OpenCPI_Platform_Development_Guide.fodt

+++ tools/ocpigen/src/ocpigen_main.cxx
@@ -330,9 +330,9 @@ main(int argc, const char **argv) {
             const char* const property[] = {"PackageName", "PackagePrefix", "Package",
              "ProjectDependencies", "Libraries", "IncludeDirs", "XmlIncludeDirs",
              "ComponentLibraries", "ProjectName", "ProjectPackage"};
-            int last = sizeof(property)/sizeof(property[0]);
+            size_t last = sizeof(property)/sizeof(property[0]);
             char *value;
-            for (int idx = 0; idx < last; ++idx) {
+            for (size_t idx = 0; idx < last; ++idx) {
               value = (char *)ezxml_cattr(xml, property[idx]);
               if (value)
                 printf("%s=%s\n", property[idx], value);
